+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Add ToString() method to voice, messaging and fax responses to generate the TwiM
 - **New Resource:** Call Recordings
 - **New Resource:** Conference Recordings
 - **New Resource:** Call Feedback Summary
+- **New Resource:** Call Feedback
 
 ## Lookups